requirements.txt
tensorflow==2.6.0
streamlit==1.0
numpy
pillow
